move AttachmentMethods, GalleryMethods and related out of Proposals to use it in Budgets
