+++ decidim-admin/app/assets/javascripts/decidim/admin/gallery.js.es6
@@ -1,0 +1,5 @@
+$(() => {
+  $(".gallery__container").on("closed.zf.callout", (event) => {
+    $(event.target).remove();
+  });
+});
