+++ decidim-budgets/app/views/decidim/budgets/admin/projects/_form.html.erb
@@ -35,26 +35,6 @@
       <% end %>
     </div>
 
-    <div class="row column gallery__container">
-      <fieldset>
-        <legend><%= t("decidim.proposals.admin.proposals.form.gallery_legend") %></legend>
-
-        <% if @form.photos.any? %>
-          <% @form.photos.each do |photo| %>
-            <div class="callout gallery__item" data-closable>
-              <%= image_tag photo.thumbnail_url, class: "thumbnail", alt: photo.file.file.filename %>
-              <%= form.hidden_field :photos, multiple: true, value: photo.id, id: "photo-#{photo.id}" %>
-              <button class="close-button" aria-label="Close" type="button" data-close>
-                <span aria-hidden="true">&times;</span>
-              </button>
-            </div>
-          <% end %>
-        <% end %>
-
-        <div class="row column">
-          <%= form.file_field :add_photos, multiple: false, label: t("decidim.proposals.admin.proposals.form.add_images") %>
-        </div>
-      </fieldset>
-    </div>
+    <%= render partial: "decidim/admin/shared/gallery", locals: { form: form } %>
   </div>
 </div>
