+++ decidim-proposals/app/assets/javascripts/decidim/proposals/admin/proposals_form.js.es6
@@ -20,10 +20,4 @@ $(() => {
     toggleDisabledHiddenFields();
 
   }
-
-  // has to be moved to core when moving gallery_methods
-  $(document).on("closed.zf.callout", (event) => {
-    $(event.target).remove();
-  });
-
 });
