+++ decidim-proposals/app/commands/decidim/proposals/update_proposal.rb
@@ -4,7 +4,7 @@ module Decidim
   module Proposals
     # A command with all the business logic when a user updates a proposal.
     class UpdateProposal < Rectify::Command
-      include AttachmentMethods
+      include Decidim::AttachmentMethods
       include HashtagsMethods
 
       # Public: Initializes the command.
